+++ core/src/com/cyphercove/covetools/assets/AssetContainer.java
@@ -15,8 +15,6 @@
  ******************************************************************************/
 package com.cyphercove.covetools.assets;
 
-import org.jetbrains.annotations.Nullable;
-
 /**
  * Optional interface for an AssetContainer, providing a loading callback and other functionality for {@link AssignmentAssetManager}.
  */
@@ -25,7 +23,7 @@ public interface AssetContainer {
      * @return A path string that will prepended to all asset paths in this container. If a directory, a trailing slash must be included.
      * May return null to prepend nothing.
      */
-    @Nullable String getAssetPathPrefix ();
+    String getAssetPathPrefix ();
 
     /**
      * Called when the AssetManager has finished loading all assets and populated the annotated asset fields in this container.