+++ core/src/com/cyphercove/covetools/utils/JsonFieldUpdater.java
@@ -26,8 +26,6 @@ import com.badlogic.gdx.utils.reflect.ReflectionException;
 import net.objecthunter.exp4j.Expression;
 import net.objecthunter.exp4j.ExpressionBuilder;
 
-import org.jetbrains.annotations.NotNull;
-
 import java.io.StringWriter;
 import java.security.AccessControlException;
 import java.util.Locale;
@@ -56,19 +54,20 @@ import java.util.Locale;
  *
  * @author cypherdare
  */
+@SuppressWarnings("rawtypes") // Raw types are necessary to work with the Json class.
 public class JsonFieldUpdater {
 
     private static final String TAG = JsonFieldUpdater.class.getName();
 
     private boolean throwErrors;
     private Json json;
-    private ObjectMap<Class, OrderedMap<String, Field>> staticFields = new ObjectMap<Class, OrderedMap<String, Field>>();
+    private ObjectMap<Class<?>, OrderedMap<String, Field>> staticFields = new ObjectMap<Class<?>, OrderedMap<String, Field>>();
 
     public JsonFieldUpdater() {
         this(new Json());
     }
 
-    public JsonFieldUpdater(@NotNull Json json) {
+    public JsonFieldUpdater(Json json) {
         this.json = json;
         json.setIgnoreUnknownFields(true);
 
@@ -107,7 +106,6 @@ public class JsonFieldUpdater {
         });
     }
 
-    @NotNull
     public Json getJson() {
         return json;
     }
@@ -120,7 +118,7 @@ public class JsonFieldUpdater {
         this.throwErrors = throwErrors;
     }
 
-    public void readFieldsToObjects(@NotNull FileHandle jsonFile, @NotNull Object... objects) {
+    public void readFieldsToObjects(FileHandle jsonFile, Object... objects) {
         readFieldsToObjects(jsonFile.readString(), objects);
     }
 
@@ -134,10 +132,10 @@ public class JsonFieldUpdater {
      * @param objects  A set of classes that can be referenced by simple name in the JSON instead of fully qualified names.
      *                 If a given object is a class, static fields will be written instead of object fields.
      */
-    public void readFieldsToObjects(@NotNull String jsonText, @NotNull Object... objects) {
+    public void readFieldsToObjects(String jsonText, Object... objects) {
         ObjectMap<String, Class> tagsToClasses = new ObjectMap<String, Class>(objects.length);
         for (Object object : objects) {
-            Class type = object instanceof Class ? (Class) object : object.getClass();
+            Class<?> type = object instanceof Class ? (Class<?>) object : object.getClass();
             tagsToClasses.put(type.getSimpleName(), type);
         }
 
@@ -152,7 +150,7 @@ public class JsonFieldUpdater {
         outer:
         for (Object object : objects) {
             if (object instanceof Class) {
-                Class type = (Class) object;
+                Class<?> type = (Class<?>) object;
                 for (JsonValue child : root) {
                     if (getClass(child, tagsToClasses) == type) {
                         readStaticFields(type, child);
@@ -240,7 +238,7 @@ public class JsonFieldUpdater {
         return nameToField;
     }
 
-    public <T> T getStaticFieldValue(@NotNull Class<T> type, @NotNull String fieldName) {
+    public <T> T getStaticFieldValue(Class<T> type, String fieldName) {
         OrderedMap<String, Field> fields = getStaticFields(type);
         Field field = fields.get(fieldName, null);
         Throwable throwable = null;
@@ -270,8 +268,7 @@ public class JsonFieldUpdater {
      * @param type The class to read.
      * @return A json representation of the class's static fields
      */
-    @NotNull
-    public static String toJson (@NotNull Class type){
+    public static String toJson (Class type){
         Field[] fields = ClassReflection.getDeclaredFields(type);
         JsonWriter writer = new JsonWriter(new StringWriter());
         writer.setOutputType(JsonWriter.OutputType.javascript);
@@ -299,9 +296,8 @@ public class JsonFieldUpdater {
      * @param object The object to read.
      * @return A json representation of the object's member fields
      */
-    @NotNull
-    public static String toJson (@NotNull Object object){
-        Class type = object.getClass();
+    public static String toJson (Object object){
+        Class<?> type = object.getClass();
         Field[] fields = ClassReflection.getDeclaredFields(type);
         JsonWriter writer = new JsonWriter(new StringWriter());
         writer.setOutputType(JsonWriter.OutputType.javascript);