+++ CHANGES.md
@@ -1,6 +1,13 @@
-# 1.2.3 (pending next release)
+# 1.2.3
  - Update to LibGDX 1.9.11
- - Deprecated Anisotropy, as its features are now available from Texture directly.
+ - Deprecated `Anisotropy`, as its features are now available from Texture directly.
+ - **[Breaking (imports only)]** WallpaperEventListener is no longer an inner interface of LiveWallpaperWrapper.
+ - Added WallpaperEventAdapter, empty implementation of WallpaperEventListener.
+ - Added LiveWallpaperAdapter, empty implementation of LiveWallpaperListener.
+ - Fleshed out CoveTools.gwt.xml, for using CoveTools with GWT.
+ - Tweens removed. See [gdx-tween](https://github.com/CypherCove/gdx-tween).
+ - **[Breaking]** Change raw return type of `LiveWallpaperInfoActivity.getWallpaperServiceClass()` to `Class<? extends Service>`.
+ - Deprecated `ColorUtil.blendRGBWithSaturation(...)` as there are some cases where it produces sudden jumps in interpolation.
 
 # 1.2.2
  - Add TextureAtlasCacher.